Move helper functions from `index.ts` to `misc.ts`
